# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAText
  description: |
    This OpenAPI document describes APIs for WASAText project.

    Studente: Francesco Zompanti
    Matricola: 2012601
    e-mail: zompanti.2012601@studenti.uniroma1.it
  version: 0.0.1
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    CookieAuth:
      type: apiKey
      in: cookie
      name: wasatext_session
  schemas:
...
    User:
      type: object
      description: User schema.
      required: [id, name, photo]
      properties:
        id:
          type: string
          description: Unique identifier of the user.
          example: "justchecco"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the user.
          example: "Francesco"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 3
          maxLength: 16
        photo:
          type: string
          description: User photo in base64 encoding.
          example: "aGPsaG8"
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 1000000
...
    Message:
      type: object
      description: A message sent in a conversation.
      required:
        - id
        - senderId
        - senderName
        - content
        - timestamp
        - attachment
        - reactionCount
        - reactingUserIds
      properties:
        id:
          type: string
          description: Unique identifier of the message.
          example: "msg123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        senderId:
          type: string
          description: ID of the user who sent the message.
          example: "justchecco2"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        senderName:
          type: string
          description: Name of the sender.
          example: "Alice"
          pattern: '^[a-zA-Z0-9 ]+$'
          minLength: 1
          maxLength: 50
        content:
          type: string
          description: Content of the message.
          example: "Hello, world!"
          pattern: '^.*$'
          minLength: 1
          maxLength: 1000
        timestamp:
          type: string
          format: date-time
          description: When the message was sent.
          example: "2025-11-01T11:08:00Z"
          minLength: 20
          maxLength: 29
        attachment:
          type: string
          description: Base64-encoded attachment (if any).
          example: ""
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 10485760
        reactionCount:
          type: integer
          format: int32
          description: Number of reactions on the message.
          example: 2
          minimum: 0
          maximum: 2147483647
        reactingUserIds:
          type: array
          description: List of user IDs who have reacted to the message.
          minItems: 0
          maxItems: 1000
          items:
            type: string
            example: "justchecco"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        replyTo:
          type: string
          description: (Optional) ID of the message being replied to.
          example: ""
          pattern: '^[a-zA-Z0-9_]*$'
          minLength: 0
          maxLength: 50
        replyContent:
          type: string
          description: (Optional) A preview of the message being replied to.
          example: ""
          pattern: '^.*$'
          minLength: 0
          maxLength: 1000
        replySenderName:
          type: string
          description: (Optional) Name of the sender of the replied-to message.
          example: ""
          pattern: '^[a-zA-Z0-9 ]*$'
          minLength: 0
          maxLength: 50
        replyAttachment:
          type: string
          description: (Optional) Base64-encoded attachment from the replied-to message.
          example: ""
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 10485760
        status:
          type: string
          description: (Optional) Delivery/read status; provided only for messages not sent by the authenticated user.
          example: "✓"
          pattern: '^.*$'
          minLength: 0
          maxLength: 10
    ForwardMessageRequest:
      type: object
      description: Request schema for forwarding a message. Either targetConversationId or targetUserId is required.
      properties:
        targetConversationId:
          type: string
          description: ID of the conversation to which the message will be forwarded.
          example: "conv456"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        targetUserId:
          type: string
          description: ID of the user to start or reuse a direct conversation with and forward the message.
          example: "user789"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        forwarderName:
          type: string
          description: (Optional) Name of the user forwarding the message.
          example: "Francesco"
          pattern: '^[a-zA-Z0-9 ]*$'
          minLength: 0
          maxLength: 50
...
    AddGroupMemberRequest:
      type: object
      description: Request schema for adding a member to a group.
      required: [userId]
      properties:
        userId:
          type: string
          description: ID of the user to add.
          example: "justchecco"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
    Group:
      type: object
      description: Group schema.
      required: [id, name, members]
      properties:
        id:
          type: string
          description: Unique identifier of the group.
          example: "group123"
          pattern: '^[a-zA-Z0-9_]+$'
          minLength: 1
          maxLength: 50
        name:
          type: string
          description: Name of the group.
          example: "Group Chat"
          pattern: '^[a-zA-Z0-9_ ]+$'
          minLength: 3
          maxLength: 50
        members:
          type: array
          description: List of user IDs that are members of the group.
          minItems: 1
          maxItems: 1000
          items:
            type: string
            description: User ID.
            example: "justchecco"
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        groupPhoto:
          type: string
          description: Base64-encoded group photo (if any).
          example: "aGPsaG8"
          pattern: '^[A-Za-z0-9+/]*={0,2}$'
          minLength: 0
          maxLength: 1000000
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
  /session:
    post:
      security: []
      tags: ["login"]
      summary: Logs in the user
      description: Authenticates an existing user and rotates the opaque session token.
      operationId: doLogin
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Login credentials.
              required: [name, password]
              properties:
                name:
                  type: string
                  description: Name to assign to the user.
                  example: Francesco
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
                password:
                  type: string
                  format: password
                  minLength: 10
                  maxLength: 128
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Login response payload.
                required: [user]
                properties:
                  user:
                    $ref: '#/components/schemas/User'
              example:
                user:
                  id: "abcdef012345"
                  name: "Francesco"
                  photo: ""
    delete:
      tags: [login]
      summary: Logs out the current user
      description: Revokes the current token and removes the browser session cookie.
      operationId: logout
      responses:
        '204':
          description: Session revoked
  /users:
    post:
      tags: [login]
      summary: Registers a new user
      description: Creates a password-protected account and returns its first opaque session token.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password]
              properties:
                name:
                  type: string
                  pattern: '^[a-zA-Z0-9_]+$'
                  minLength: 3
                  maxLength: 16
                password:
                  type: string
                  format: password
                  minLength: 10
                  maxLength: 128
      responses:
        '201':
          description: User registered
        '409':
          description: Username already exists
...
  /conversations/{conversationId}/message/{messageId}/forward:
    post:
      tags: [message]
      summary: Forwards an existing message to another conversation
      description: Forwards a message to another conversation
      operationId: forwardMessage
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: conversationId
          in: path
          required: true
          description: Conversation ID with the message to forward
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        - name: messageId
          in: path
          required: true
          description: Message ID to forward
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      requestBody:
        description: JSON payload with the target conversation ID
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardMessageRequest'
            example:
              targetConversationId: "conversazionesempio"
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: "msg123"
                senderId: "justchecco"
                senderName: "Francesco"
                content: "Ciao"
                timestamp: "2025-11-01T11:08:00Z"
                attachment: ""
                reactionCount: 0
                reactingUserIds: []
  /conversations/{conversationId}/message/{messageId}:
    delete:
      tags: [message]
      summary: Deletes a message
      description: Deletes a message from the conversation
      operationId: deleteMessage
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: conversationId
          in: path
          required: true
          description: Conversation ID
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
        - name: messageId
          in: path
          required: true
          description: Message ID to delete
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '204':
          description: Message deleted successfully
...
  /search:
    get:
      tags: [conversation, user]
      summary: Searches for users by username
      description: Searches for users matching a query string.
      operationId: searchUsers
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: username
          in: query
          required: true
          description: Username query parameter
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '200':
          description: Users matching the query.
          content:
            application/json:
              schema:
                type: object
                description: Search result payload containing the matched users.
                required: [users]
                properties:
                  users:
                    type: array
                    description: Array of User objects.
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/User'
              example:
                users:
                  - id: "justchecco"
                    name: "Francesco"
                    photo: "aGPsaG8="
...
  /groups/{groupId}:
    get:
      tags: [group]
      summary: Retrieves details of a specific group
      description: Retrieves group details including members and group photo.
      operationId: getGroup
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group.
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '200':
          description: Group details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              example:
                id: "group123"
                name: "Group Chat"
                members: ["justchecco","justchecco2"]
                groupPhoto: "aGPsaG8"
    delete:
      tags: [group]
      summary: Leaves a group
      description: Removes the authenticated user from the specified group
      operationId: leaveGroup
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group to leave
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      responses:
        '204':
          description: Left group successfully
    post:
      tags: [group]
      summary: Adds a user to a group
      description: Adds a user to the specified group
      operationId: addToGroup
      security: [ { BearerAuth: [] } ]
      parameters:
        - name: groupId
          in: path
          required: true
          description: ID of the group
          schema:
            type: string
            pattern: '^[a-zA-Z0-9_]+$'
            minLength: 1
            maxLength: 50
      requestBody:
        description: JSON payload with the user ID to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberRequest'
      responses:
        '204':
          description: User added to group successfully
...